DiffEnhancer/requirements.txt
diffusers==0.21.4
accelerate
transformers
xformers
ultralytics
webdataset
open_clip_torch
salesforce-lavis==1.0.2
pytorch_lightning
einops
gradio<4.0.0
numpy>=1.17
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
matplotlib